+++ HISTORY.rst
@@ -2,6 +2,10 @@
 History
 =======
 
+0.9.13 (2018-02-11)
+------------------
+* short original links in captions
+
 0.9.12 (2018-02-11)
 ------------------
 * flood is now source